Naive encoding
This simply tries decodeUtf8' then fallsback to decodeLatin1 if that
fails. For all data we've seen so far this will work because we have a
mixture of utf8 and latin1 but all bytes are valid latin1 and so we
could still be storing wrongly encoded data if any other encodings crop
up.